refactor: extract dxf leaf utility modules
