+++ docs/DXF_BATCH1_LEAF_MODULES_VERIFICATION.md
@@ -1,0 +1,36 @@
+# DXF Batch 1 Leaf Modules Verification
+
+## Clean Worktree Build
+
+Built successfully in:
+
+- `build-codex`
+
+Targets built:
+
+- `cadgf_dxf_importer_plugin`
+- `convert_cli`
+- DXF/DWG tool test executables needed for clean validation
+
+## Clean Worktree Validation
+
+Observed in the clean worktree:
+
+- `22/26` runnable `dxf|dwg` tests passed after excluding two known baseline blockers:
+  - `test_dxf_leader_metadata_run` is blocked by a pre-existing compile break in `test_dxf_leader_metadata.cpp` against `core_c_api.h`
+  - `convert_cli_dxf_style_smoke` is blocked by missing `cmake/RunConvertCliDxfLineStyle.cmake`
+
+The remaining 4 failing runtime tests are pre-existing on `origin/main` and reproduced unchanged there:
+
+- `test_dxf_multi_layout_metadata_run`
+- `test_dxf_paperspace_insert_styles_run`
+- `test_dxf_paperspace_insert_dimension_run`
+- `test_dxf_paperspace_combo_run`
+
+## Acceptance Statement
+
+This packet does not introduce a new clean-baseline failure surface relative to `origin/main`.
+
+## Hygiene
+
+- `git diff --check` passed
